Move actions to new package
This allows the package to be imported into a magefile
